+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for octet-stream content type (#116)
 - Support for [nullable](https://swagger.io/docs/specification/data-models/data-types/#null) (#99)
 - Union properties defined using oneOf (#98)
+- Added support for lists of strings, integers, floats and booleans (#165). Thanks @Maistho!
 
 
 ## 0.5.3 - 2020-08-13